pyproject.toml
[tool.poetry]
name = "spotlight"
version = "0.1.0"
description = ""
authors = ["Roma"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
djangorestframework = "^3.14.0"
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.2.0"
django-rest-multiple-models = "^2.1.3"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
